refactor: move parquet-format specific functionality to the parquet submodule, leaving only the config options in the config module.
